requirements/base.txt
Django
django-cleanup
Pillow
django-environ
